# openapi.yaml
openapi: 3.0.3
info:
  title: AI数字名片 API
  description: AI数字名片后端服务 - 模块化架构
  version: 2.0.0
  contact:
    name: AI数字名片团队
servers:
- url: http://localhost:8201
  description: 本地开发
- url: https://api.example.com
  description: 生产环境
paths:
...
  /api/brochures:
    post:
      summary: 创建画册
      tags:
      - 画册
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureCreate'
    get:
      summary: 获取画册列表
      tags:
      - 画册
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: integer
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: 画册列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrochureResponse'
  /api/brochures/{brochure_id}:
    get:
      summary: 获取画册详情
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
    put:
      summary: 更新画册
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrochureResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureUpdate'
    delete:
      summary: 删除画册
      tags:
      - 画册
      parameters:
      - name: brochure_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 删除成功
        '204':
          description: 无内容
      security:
      - BearerAuth: []
...
  /u/{username}:
    get:
      summary: 公开用户Profile页面(SSR)
      tags:
      - 公开页面
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: HTML Profile页面
...
  /api/export/json:
    get:
      summary: 导出名片数据(JSON)
      tags:
      - 数据导出
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
  /api/export/csv:
    get:
      summary: 导出名片数据(CSV)
      tags:
      - 数据导出
      parameters: []
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌, 通过 /api/auth/login 获取
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key, 通过 /api/api-keys 创建
  schemas:
...
    BrochureCreate:
      type: object
      required:
      - title
      - pages
      properties:
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
          enum:
          - partner
          - client
          - investor
          - supplier
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
    BrochureUpdate:
      type: object
      properties:
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
    BrochureResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        cover:
          type: string
        purpose:
          type: string
        status:
          type: string
        share_token:
          type: string
        view_count:
          type: integer
        pages_count:
          type: integer
        album_meta:
          type: object
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PageSchema:
      type: object
      properties:
        id:
          type: integer
        sort_order:
          type: integer
        content_type:
          type: string
        content:
          type: string
        image_url:
          type: string
        media_url:
          type: string
        ai_summary:
          type: string
...